refactor コンポーネントに切り分け
